requirements.txt
glfw
PyOpenGL
numpy
Pillow
plyfile
PyGLM
